fix: harden chart identity and lifecycle
- rebuild time-derived state from explicit sequence generations and reject stale series/pane handles
- explicitly dispose WASM charts and extensions while broadcasting shared WebGPU loss to every live chart
- expose structured ingestion diagnostics and make portable browser validation publication-blocking

Verification: cargo fmt/clippy/test, wasm clippy, GPUI backend tests, package lint/build/typecheck/pack, SSR import, 124 Playwright tests, benchmark test/smoke, and release perf gate.

+++ crates/nucleuscharts_engine/src/frame/axis.rs
@@ -1031,7 +1031,7 @@ impl ChartEngine {
                     price_text: show_price.then_some(text),
                     title,
                     countdown,
-                    group_id: series.id as u32,
+                    group_id: series.id,
                     // The price row stays centered on the value coordinate; the countdown row
                     // hangs below, so a FULL cluster's center shifts down by half the countdown
                     // row. Without a price row the cluster centers on the value directly.
@@ -1077,7 +1077,7 @@ impl ChartEngine {
                             price_text: Some(quote_text),
                             title: Some(side.to_string()),
                             countdown: None,
-                            group_id: (1u32 << 30) + series.id as u32 * 4 + side_offset,
+                            group_id: (1u32 << 30) + series.id * 4 + side_offset,
                             y: quote_y,
                             height: row_height,
                             top_height: row_height,
